requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.7.2
boto3==1.29.0
botocore==1.32.0
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
gunicorn==21.2.0
jmespath==1.0.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.18
uuid==1.30
